MOBILE MENU COMPLETE AND LINTERS ERR FIX
