--- src/components/hamburger.css
@@ -1,3 +1,0 @@
-/* .sidebar-container {
-  display: none;
-} */
